+++ extensions/Worker.Extensions.EventGrid/release_notes.md
@@ -4,6 +4,6 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker.Extensions.EventGrid <version>
+### Microsoft.Azure.Functions.Worker.Extensions.EventGrid 3.4.1
 
-- <entry>
+- Updated `Microsoft.Azure.WebJobs.Extensions.EventGrid` reference to 3.3.1